+++ R/RKorAPClient-package.R
@@ -1,0 +1,15 @@
+#' @references
+#' Kupietz, Marc / Diewald, Nils / Margaretha, Eliza (2020):
+#' RKorAPClient: An R package for accessing the German Reference Corpus DeReKo
+#' via KorAP. In: Calzolari, Nicoletta, Frédéric Béchet, Philippe Blache,
+#' Khalid Choukri, Christopher Cieri,  Thierry Declerck, Sara Goggi,
+#' Hitoshi Isahara, Bente Maegaard, Joseph Mariani, Hélène Mazo,
+#' Asuncion Moreno, Jan Odijk, Stelios Piperidis (eds.):
+#' Proceedings of The 12th Language Resources and Evaluation Conference (LREC 2020)
+#' Marseille: European Language Resources Association (ELRA), 7017-7023.
+#' \url{http://www.lrec-conf.org/proceedings/lrec2020/pdf/2020.lrec-1.867.pdf}
+#'
+#' @keywords internal
+"_PACKAGE"
+#' [1] "_PACKAGE"
+#'